Decompouse Layout component into smaller components
The Layout component was getting too complex and hard to manage, so it was broken down into smaller, more manageable components.
Also a new component ProtectedRoutesWrapper was added that takes as children the routes that should be private and checks if the user is authorized. If he is authorized then the route is accessible otherwise the user is redirected to the login page